Created a function get_breach_count
The function takes two parameters, hash_to_check and hashes_list. The
first parameter is the hashed version of user's password and second is
the list of all the hashed passwords returned by api.

Function checks if user's hashed password has been in any data breach.
If yes, it returns the number of times it has been in data breach.